requirements.txt
langchain
langchain-openai
langchain-community
langchain-ollama
langchain-core
google-search-results
beautifulsoup4
pyyaml
python-dotenv
langchain-experimental>=0.0.60
 
# Neo4j GraphRAG
neo4j>=5.14.0
langchain-neo4j>=0.1.0
qdrant-client>=1.11.0
 
# PDF processing
pypdf>=4.0.0
pdfplumber>=0.11.0
unstructured[pdf,docx]
 
# Embeddings & LLM
openai>=1.30.0
tiktoken>=0.7.0
 
# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0

# Django Backend
django>=5.0.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0

# Agent Veille
httpx>=0.27.0

# Green Analysis Module (LangGraph multi-agent pipeline)
langgraph>=1.0.10
langgraph-checkpoint-sqlite
langchain-groq
langchain-huggingface
faiss-cpu>=1.13.2
sentence-transformers
tavily-python
pydantic-settings
